feat: add typescript
Migrate all files and scripts to typescript
