+++ .eslintrc.json
@@ -1,17 +1,21 @@
 {
   "env": {
-    "node": true,
     "browser": true,
-    "commonjs": true,
-    "es2021": true
+    "es2021": true,
+    "node": true
   },
-  "extends": ["eslint:recommended", "plugin:storybook/recommended", "prettier"],
+  "extends": [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:storybook/recommended",
+    "prettier"
+  ],
   "overrides": [],
+  "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "ecmaVersion": "latest",
     "sourceType": "module"
   },
-  "rules": {
-    "no-unused-vars": "warn"
-  }
+  "plugins": ["@typescript-eslint"],
+  "rules": {}
 }
