+++ .storybook/main.ts
@@ -1,0 +1,23 @@
+import type { StorybookConfig } from '@storybook/html-vite';
+
+const config: StorybookConfig = {
+  stories: [
+    '../storybook/stories/**/*.mdx',
+    '../storybook/stories/**/*.stories.@(js|jsx|ts|tsx)',
+  ],
+  addons: [
+    '@storybook/addon-links',
+    '@storybook/addon-essentials',
+    '@storybook/addon-interactions',
+  ],
+  staticDirs: ['../node_modules/@fontsource/montserrat'],
+  framework: {
+    name: '@storybook/html-vite',
+    options: {},
+  },
+  docs: {
+    autodocs: true,
+  },
+};
+
+export default config;
