--- .storybook/preview.js
@@ -1,11 +1,0 @@
-import '../packages/storybook/tailwind.css';
-
-export const parameters = {
-  actions: { argTypesRegex: '^on[A-Z].*' },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
-};
