+++ .storybook/preview.ts
@@ -1,0 +1,11 @@
+import '../storybook/tailwind.css';
+
+export const parameters = {
+  actions: { argTypesRegex: '^on[A-Z].*' },
+  controls: {
+    matchers: {
+      color: /(background|color)$/i,
+      date: /Date$/,
+    },
+  },
+};
