+++ package.json
@@ -2,11 +2,13 @@
   "name": "@bkbank/ui",
   "version": "1.2.0-beta.1",
   "description": "The Bk Bank Design System written as a TailwindCSS Plugin",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
-    "build": "run-p build:**",
-    "build:base": "postcss --config packages/core packages/core/base/index.css -o dist/base.css",
-    "build:components": "postcss --config packages/core packages/core/components/index.css -o dist/components.css",
+    "build": "npm-run-all tsc --parallel build:**",
+    "tsc": "tsc",
+    "build:all": "postcss packages/core/all.css -o dist/all.css",
+    "build:base": "postcss packages/core/base.css -o dist/base.css",
+    "build:components": "postcss packages/core/components.css -o dist/components.css",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
     "semantic-release": "semantic-release"
@@ -36,6 +38,9 @@
     "@storybook/html-vite": "7.0.0-beta.54",
     "@storybook/testing-library": "0.0.14-next.1",
     "@tailwindcss/forms": "^0.5.3",
+    "@types/postcss-js": "^4.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.53.0",
+    "@typescript-eslint/parser": "^5.53.0",
     "autoprefixer": "^10.4.13",
     "clsx": "^1.2.1",
     "eslint": "^8.34.0",
