+++ packages/core/all.css
@@ -1,0 +1,2 @@
+@import './base.css';
+@import './components.css';
