+++ packages/core/components.css
@@ -1,0 +1,9 @@
+@import './components/avatar.css';
+@import './components/badge.css';
+@import './components/breadcrumb.css';
+@import './components/button.css';
+@import './components/checkbox.css';
+@import './components/input.css';
+@import './components/input-group.css';
+@import './components/radio.css';
+@import './components/switch.css';
