--- packages/core/components/index.css
@@ -1,9 +1,0 @@
-@import './avatar.css';
-@import './badge.css';
-@import './breadcrumb.css';
-@import './button.css';
-@import './checkbox.css';
-@import './input.css';
-@import './input-group.css';
-@import './radio.css';
-@import './switch.css';
