--- packages/storybook/stories/components/button/create-button.js
@@ -1,44 +1,0 @@
-import clsx from 'clsx';
-
-const colors = {
-  primary: 'btn-primary',
-  secondary: 'btn-secondary',
-};
-
-const variants = {
-  solid: 'btn-solid',
-  outline: 'btn-outline',
-  ghost: 'btn-ghost',
-};
-
-const sizes = {
-  small: '',
-  medium: 'btn-md',
-  large: 'btn-lg',
-};
-
-export function createButton({
-  size = 'small',
-  variant = 'solid',
-  color = 'primary',
-  disabled = false,
-  withIcon = false,
-}) {
-  return `
-    <button class="${clsx(
-      'btn',
-      variants[variant],
-      colors[color],
-      sizes[size],
-    )}" type="button"${disabled ? ' disabled' : ''}>
-      ${
-        withIcon
-          ? `<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" width="1em" hight="1em">
-        <path stroke-linecap="round" stroke-linejoin="round" d="M15.75 10.5V6a3.75 3.75 0 10-7.5 0v4.5m11.356-1.993l1.263 12c.07.665-.45 1.243-1.119 1.243H4.25a1.125 1.125 0 01-1.12-1.243l1.264-12A1.125 1.125 0 015.513 7.5h12.974c.576 0 1.059.435 1.119 1.007zM8.625 10.5a.375.375 0 11-.75 0 .375.375 0 01.75 0zm7.5 0a.375.375 0 11-.75 0 .375.375 0 01.75 0z" />
-      </svg>`
-          : ''
-      }
-      Shop now
-    </button>
-  `;
-}
