--- packages/storybook/tailwind.css
@@ -1,11 +1,0 @@
-@import 'tailwindcss/base';
-@import '../core/base/index.css';
-@import 'tailwindcss/components';
-@import '../core/components/index.css';
-@import 'tailwindcss/utilities';
-
-@import '@fontsource/montserrat/variable.css';
-
-html {
-  @apply font-montserrat;
-}
